requirements.txt
# =============================
# War-Room Bot — Dependencies
# =============================

# Agent Framework
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0

# Market Data
finnhub-python>=2.4.19

# News & Search
tavily-python>=0.5.0
xgboost>=2.0.3
statsmodels>=0.14.1
scipy>=1.12.0

# Vector Database
pinecone>=3.0.0

# Broker
alpaca-py>=0.21.0

# Technical Analysis & Math
numpy>=1.26.0
pandas>=2.1.0

# Dashboard & Logging
rich>=13.7.0

# Configuration
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
